refactor: EyeIcon, ClearInputIcon 분리 작성
